+++ src/app/clients/getClientsTableData.ts
@@ -1,6 +1,6 @@
+import { CongestionChargeDetails, ProcessingData } from "@/app/clients/fetchDataFromServer";
 import { Schema } from "@/database_utils";
 import { Datum } from "@/components/Tables/Table";
-import supabase from "@/supabaseServer";
 import {
     familyCountToFamilyCategory,
     formatDatetimeAsDate,
@@ -21,50 +21,17 @@ export interface ClientsTableRow extends Datum {
     lastStatus: string;
 }
 
-export type ProcessingData = Awaited<ReturnType<typeof getProcessingData>>;
-
-// eslint-disable-next-line @typescript-eslint/explicit-function-return-type
-const getProcessingData = async () => {
-    const response = await supabase
-        .from("parcels")
-        .select(
-            `
-        parcel_id:primary_key,
-        collection_centre,
-        collection_datetime,
-        packing_datetime,
-        
-        client:clients (
-            primary_key,
-            full_name,
-            address_postcode,
-            flagged_for_attention,
-            signposting_call_required,
-            
-            family:families (
-                age,
-                gender
-            )
-        ),
-        
-        events (
-            event_name,
-            timestamp
-        )
-    `
-        )
-        .order("packing_datetime", { ascending: false })
-        .order("timestamp", { ascending: false, foreignTable: "events" })
-        .limit(1, { foreignTable: "events" });
-
-    return response.data ?? [];
-};
-
-export const processingDataToClientsTableData = async (
-    processingData: ProcessingData
-): Promise<ClientsTableRow[]> => {
+export const processingDataToClientsTableData = (
+    processingData: ProcessingData,
+    congestionCharge: CongestionChargeDetails[]
+): ClientsTableRow[] => {
     const clientTableRows: ClientsTableRow[] = [];
-    const congestionChargeDetails = await getCongestionChargeDetails(processingData);
+
+    if (processingData.length !== congestionCharge.length) {
+        throw new Error(
+            `Invalid inputs, got length ${processingData.length} and ${congestionCharge.length}`
+        );
+    }
 
     for (let index = 0; index < processingData.length; index++) {
         const parcel = processingData[index];
@@ -79,7 +46,7 @@ export const processingDataToClientsTableData = async (
             familyCategory: familyCountToFamilyCategory(client.family.length),
             addressPostcode: client.address_postcode,
             collectionCentre: parcel.collection_centre ?? "-",
-            congestionChargeApplies: congestionChargeDetails[index].congestionCharge,
+            congestionChargeApplies: congestionCharge[index].congestionCharge,
             packingDate: formatDatetimeAsDate(parcel.packing_datetime),
             packingTimeLabel: datetimeToPackingTimeLabel(parcel.packing_datetime),
             collectionDatetime: parcel.collection_datetime ?? "-",
@@ -88,21 +55,6 @@ export const processingDataToClientsTableData = async (
     }
 
     return clientTableRows;
-};
-
-export const getCongestionChargeDetails = async (
-    processingData: ProcessingData
-): Promise<{ postcode: string; congestionCharge: boolean }[]> => {
-    const postcodes = [];
-    for (const parcel of processingData) {
-        postcodes.push(parcel.client!.address_postcode);
-    }
-
-    const response = await supabase.functions.invoke("check-congestion-charge", {
-        body: { postcodes: postcodes },
-    });
-
-    return JSON.parse(response.data);
 };
 
 export const datetimeToPackingTimeLabel = (datetime: string | null): string => {
@@ -121,9 +73,4 @@ export const eventToStatusMessage = (
     }
 
     return `${event.event_name} @ ${formatDatetimeAsDate(event.timestamp)}`;
-};
-
-export const getClientsTableData = async (): Promise<ClientsTableRow[]> => {
-    const processingData = await getProcessingData();
-    return processingDataToClientsTableData(processingData);
 };
