+++ src/app/clients/getExpandedClientDetails.ts
@@ -1,4 +1,4 @@
-import supabase from "@/supabaseClient";
+import { RawClientDetails } from "@/app/clients/fetchDataFromServer";
 import { Schema } from "@/database_utils";
 import { Data } from "@/components/DataViewer/DataViewer";
 
@@ -20,49 +20,6 @@ export const formatDatetimeAsDate = (datetime: string | null): string => {
     }
 
     return new Date(datetime).toLocaleDateString("en-GB");
-};
-
-export type RawClientDetails = Awaited<ReturnType<typeof getRawClientDetails>>;
-
-// eslint-disable-next-line @typescript-eslint/explicit-function-return-type
-export const getRawClientDetails = async (parcelId: string) => {
-    const response = await supabase
-        .from("parcels")
-        .select(
-            `
-        voucher_number,
-        packing_datetime,
-
-        client:clients(
-            primary_key,
-            full_name,
-            phone_number,
-            delivery_instructions,
-            address_1,
-            address_2,
-            address_town,
-            address_county,
-            address_postcode,
-
-            family:families(
-                age,
-                gender
-            ),
-
-            dietary_requirements,
-            feminine_products,
-            baby_food,
-            pet_food,
-            other_items,
-            extra_information
-        )
-
-    `
-        )
-        .eq("primary_key", parcelId)
-        .single();
-
-    return response.data;
 };
 
 export interface ExpandedClientDetails extends Data {
@@ -199,11 +156,4 @@ export const formatBreakdownOfChildrenFromFamilyDetails = (
     }
 
     return childDetails.join(", ");
-};
-
-export const getExpandedClientDetails = async (
-    parcelId: string
-): Promise<ExpandedClientDetails> => {
-    const rawDetails = await getRawClientDetails(parcelId);
-    return rawDataToExpandedClientDetails(rawDetails);
 };
